+++ src/state/auth/reducer.js
@@ -1,6 +1,6 @@
 import { initialState } from './state';
 import {
-  REGISTER_REQUEST, REGISTER_SUCCESS, REGISTER_FAILURE, SIGN_IN_SUCCESS, LOADING, SIGN_IN_ERROR
+  REGISTER_REQUEST, REGISTER_SUCCESS, REGISTER_FAILURE, SIGN_IN_SUCCESS, LOADING, SIGN_IN_ERROR, SOCIAL_AUTH
 } from './actionTypes';
 
 export default (state = initialState, action) => {
@@ -30,11 +30,17 @@ export default (state = initialState, action) => {
       return {
         ...state,
         signin: action.payload
+      }
+    case SOCIAL_AUTH:
+      return {
+        ...state,
+        login: action.payload
       };
     case SIGN_IN_ERROR:
       return {
         ...state,
-        signin: action.payload
+        signin: action.payload,
+        login: action.payload
       };
     default:
       return state;
