--- src/state/auth/reducers.js
@@ -1,19 +1,0 @@
-import { SOCIAL_AUTH, SIGN_IN_ERROR } from './actionTypes';
-import { initialState } from './state';
-
-export default (state = initialState, action) => {
-  switch (action.type) {
-    case SOCIAL_AUTH:
-      return {
-        ...state,
-        login: action.payload
-      };
-    case SIGN_IN_ERROR:
-      return {
-        ...state,
-        login: action.payload
-      };
-    default:
-      return state;
-  }
-};
